sidebar replacement
closes #425
