+++ package.json
@@ -44,6 +44,7 @@
     "react-lottie": "^1.2.3",
     "react-notification-alert": "0.0.13",
     "react-redux": "^7.2.2",
+    "react-router-bootstrap": "^0.25.0",
     "react-router-dom": "5.2.0",
     "react-scripts": "4.0.1",
     "react-select": "3.1.1",
