+++ src/assets/scss/spartan.scss
@@ -25,6 +25,7 @@
 @import 'spartan/custom/modal.scss';
 @import 'spartan/custom/nav.scss';
 @import 'spartan/custom/navbar.scss';
+@import 'spartan/custom/offcanvas.scss';
 @import 'spartan/custom/pagination.scss';
 @import 'spartan/custom/tabs.scss';
 @import 'spartan/custom/pills.scss';
