+++ src/assets/scss/spartan/custom/_close.scss
@@ -1,29 +1,10 @@
-.close { 
-    @if $enable-transitions {
-        transition: $transition-base;
-    } 
-    &>span:not(.sr-only) {
-        background-color: $close-bg;
-        color: $close-color;
-        line-height: 17px;
-        height: 1.25rem;
-        width: 1.25rem; 
-        border-radius: 50%;
-        font-size: 1.25rem;
-        display: block;
-        @if $enable-transitions {
-            transition: $transition-base;
-        } 
-    } 
-
-    &:hover,
-    &:focus {
-        background-color: $close-hover-bg;
-        color: $close-hover-color;
-        outline: none;
-        
-        span:not(.sr-only) {
-            background-color: $close-hover-bg;
-        }
-    }
+.btn-close {
+    filter: invert(100%) sepia(99%) saturate(2%) hue-rotate(317deg) brightness(110%) contrast(100%);
 }
+
+// LIGHT MODE
+.white-content {
+    .btn-close {
+        filter: invert(0%) sepia(99%) saturate(7500%) hue-rotate(318deg) brightness(97%) contrast(103%);
+    }
+  }
