+++ src/assets/scss/spartan/custom/_footer.scss
@@ -1,4 +1,6 @@
 .footer {
+	padding: 10px;
+	
 	.nav {
 		display: inline-block;
 		float: left;
