+++ src/assets/scss/spartan/custom/_navbar.scss
@@ -46,6 +46,7 @@
 .header {
   height: 67px; 
   margin-bottom:0;
+  border-top: 2px solid #fb2715;
 }
 
 .footer {
