+++ src/components/Header/Header.js
@@ -7,28 +7,25 @@ import AddressConn from '../Common/AddressConn'
 import ThemeSwitcher from '../Common/ThemeSwitcher'
 import Supply from '../Supply/Supply'
 import './Header.scss'
+import Sidebar from '../Sidebar/Sidebar'
 
-const Header = (props) => (
+const Header = () => (
   <Navbar className="header" fixed="top">
-    <Container>
-      {/* ADD TOGGLE FOR SIDEBAR HERE */}
-
-      <Link to="/" className="navbar-brand">
-        <SpartanLogo className="my-auto" />
-        <Navbar.Brand className="d-none d-sm-inline ms-2">
-          Spartan Protocol
-        </Navbar.Brand>
-      </Link>
+    <Container fluid>
+      <div>
+        <Sidebar />
+        <Link to="/" className="navbar-brand ms-2">
+          <SpartanLogo className="my-auto" />
+          <Navbar.Brand className="d-none d-sm-inline ms-2">
+            Spartan Protocol
+          </Navbar.Brand>
+        </Link>
+      </div>
 
       <Nav>
         <ThemeSwitcher />
         <LanguageDropdown />
-        <AddressConn
-          changeStates={props.changeStates}
-          changeNotification={props.changeNotification}
-          connectedTokens={props.connectedTokens}
-          connectingTokens={props.connectingTokens}
-        />
+        <AddressConn />
         <Supply />
       </Nav>
     </Container>
