+++ src/components/layout/Common.js
@@ -1,146 +1,35 @@
-import React, { useState, useRef, useEffect } from 'react'
-import { Route, Switch, Redirect, useLocation } from 'react-router-dom'
-import { useTranslation } from 'react-i18next'
+import React, { useRef } from 'react'
+import { Route, Switch, Redirect } from 'react-router-dom'
 import NotificationAlert from 'react-notification-alert'
 
 import Header from '../Header/Header'
 import Footer from '../Footer/Footer'
-import Sidebar from '../Sidebar/Sidebar'
 
-import routes from '../../routes'
-
-import logo from '../../assets/img/spartan_red_medium.svg'
 import DataManager from '../DataManager/DataManager'
+import { routes } from '../../routes'
 
 const Common = () => {
-  const activeColor = 'blue'
-  const { t } = useTranslation()
-  const [sidebarOpened, setSidebarOpened] = useState(false)
-  const mainPanelRef = useRef(null)
   const notificationAlertRef = useRef(null)
-  const location = useLocation()
-  useEffect(() => {
-    document.body.classList.add('sidebar-mini')
-    document.documentElement.scrollTop = 0
-    document.scrollingElement.scrollTop = 0
-    if (mainPanelRef.current) {
-      mainPanelRef.current.scrollTop = 0
-    }
-  }, [location])
 
   const getRoutes = (tempRoutes) =>
-    tempRoutes.map((prop) => {
-      if (prop.collapse) {
-        return getRoutes(prop.views)
-      }
-      // if (prop.layout === '/dapp') {
-      return (
-        <Route
-          path={prop.path}
-          component={prop.component}
-          key={prop.path + prop.name}
-        />
-      )
-      // }
-      // return null
-    })
-
-  const getActiveRoute = (tempRoutes) => {
-    const activeRoute = 'Default Brand Text'
-    for (let i = 0; i < tempRoutes.length; i++) {
-      if (tempRoutes[i].collapse) {
-        const collapseActiveRoute = getActiveRoute(tempRoutes[i].views)
-        if (collapseActiveRoute !== activeRoute) {
-          return collapseActiveRoute
-        }
-      } else if (window.location.pathname.indexOf(tempRoutes[i].path) !== -1) {
-        return tempRoutes[i].name
-      }
-    }
-    return activeRoute
-  }
-
-  const handleMiniClick = () => {
-    document.body.classList.toggle('sidebar-mini')
-    document.body.classList.toggle('no-sidebar-mini')
-  }
-
-  const toggleSidebar = () => {
-    setSidebarOpened(!sidebarOpened)
-    document.documentElement.classList.toggle('nav-open')
-  }
-
-  const closeSidebar = () => {
-    setSidebarOpened(false)
-    document.documentElement.classList.remove('nav-open')
-    document.body.classList.remove('no-sidebar-mini')
-  }
-
-  const clickOutSidebar = (e) => {
-    const sidebar = document.body.getElementsByClassName('sidebar-wrapper')[0]
-    if (e.target.className.length > 0) {
-      if (
-        !sidebar?.contains(e.target) &&
-        !e.target.className?.includes('icon-menu-open') &&
-        !e.target.className?.includes('icon-menu-closed') &&
-        !e.target.className?.includes('navbar-toggler')
-      ) {
-        setSidebarOpened(false)
-        document.body.classList.add('sidebar-mini')
-        document.body.classList.remove('no-sidebar-mini')
-        closeSidebar()
-      }
-    }
-  }
+    tempRoutes.map((prop) => (
+      <Route
+        path={prop.path}
+        component={prop.component}
+        key={prop.path + prop.name}
+      />
+    ))
 
   return (
-    <div
-      className="wrapper"
-      onClick={(e) => {
-        clickOutSidebar(e)
-      }}
-      onKeyDown={(e) => {
-        if (e.key === KeyboardEvent.Spacebar) {
-          clickOutSidebar(e)
-        }
-      }}
-      role="tree"
-      tabIndex={0}
-    >
+    <div className="wrapper">
       <div className="rna-container">
         <DataManager />
         <NotificationAlert ref={notificationAlertRef} />
       </div>
-      <div className="navbar-minimize-fixed" style={{ opacity: 0 }}>
-        <button
-          type="button"
-          className="minimize-sidebar btn btn-link btn-just-icon"
-          onClick={handleMiniClick}
-        >
-          <i className="spartan-icons icon-align-center visible-on-sidebar-regular text-muted" />
-          <i className="spartan-icons icon-bullet-list-67 visible-on-sidebar-mini text-muted" />
-        </button>
-      </div>
-      <Sidebar
-        routes={routes(t)}
-        activeColor={activeColor}
-        logo={{
-          innerLink: '/',
-          text: 'Spartan Protocol',
-          imgSrc: logo,
-        }}
-        closeSidebar={closeSidebar}
-      />
-      <div className="main-panel" ref={mainPanelRef} data={activeColor}>
-        <Header
-          location={location}
-          handleMiniClick={handleMiniClick}
-          brandText={getActiveRoute(routes(t))}
-          sidebarOpened={sidebarOpened}
-          toggleSidebar={toggleSidebar}
-        />
+      <div className="main-panel">
+        <Header />
         <Switch>
-          {getRoutes(routes(t))}
+          {getRoutes(routes)}
           <Redirect from="*" to="/home" />
         </Switch>
         <Footer />
